Refactor: split test file
